move packages in pkg
Let's use a more standard Go project layout.
