+++ pkg/mutator/testdata/fixtures/dec_go
