+++ pkg/mutator/testdata/fixtures/mul_go
