+++ pkg/report/report.go
@@ -18,8 +18,8 @@ package report
 
 import (
 	"github.com/fatih/color"
-	"github.com/go-gremlins/gremlins/log"
-	"github.com/go-gremlins/gremlins/mutant"
+	"github.com/go-gremlins/gremlins/pkg/log"
+	"github.com/go-gremlins/gremlins/pkg/mutant"
 	"github.com/hako/durafmt"
 	"time"
 )
